+++ src/main/java/com/mraof/minestuck/block/machine/ComputerBlock.java
@@ -1,12 +1,11 @@
 package com.mraof.minestuck.block.machine;
 
 import com.mraof.minestuck.block.MSBlockShapes;
-import com.mraof.minestuck.block.MSBlocks;
 import com.mraof.minestuck.block.MSProperties;
 import com.mraof.minestuck.blockentity.ComputerBlockEntity;
 import com.mraof.minestuck.client.gui.MSScreenFactories;
 import com.mraof.minestuck.computer.ProgramData;
-import com.mraof.minestuck.computer.ComputerThemes;
+import com.mraof.minestuck.computer.theme.ComputerTheme;
 import com.mraof.minestuck.item.MSItems;
 import com.mraof.minestuck.player.IdentifierHandler;
 import com.mraof.minestuck.skaianet.client.SkaiaClient;
@@ -47,12 +46,20 @@ public class ComputerBlock extends MachineBlock implements EntityBlock
 	
 	public final Map<Direction, VoxelShape> shapeOn, shapeOff;
 	
+	public final String computerTheme;
+	
 	public ComputerBlock(Map<Direction, VoxelShape> shapeOn, Map<Direction, VoxelShape> shapeOff, Properties properties)
+	{
+		this(shapeOn, shapeOff, ComputerTheme.DEFAULT_NAME, properties);
+	}
+	
+	public ComputerBlock(Map<Direction, VoxelShape> shapeOn, Map<Direction, VoxelShape> shapeOff, String computerTheme, Properties properties)
 	{
 		super(properties);
 		registerDefaultState(defaultBlockState().setValue(STATE, State.OFF));
 		this.shapeOn = shapeOn;
 		this.shapeOff = shapeOff;
+		this.computerTheme = computerTheme;
 	}
 	
 	@Override
@@ -108,8 +115,7 @@ public class ComputerBlock extends MachineBlock implements EntityBlock
 			{
 				computer.owner = IdentifierHandler.encode(player);
 				
-				if(newState.is(MSBlocks.OLD_COMPUTER.get()))
-					computer.setTheme("sburb_95");
+				computer.setTheme(computerTheme);
 			}
 			
 			newState.use(level, player, handIn, hit);