+++ src/main/java/com/mraof/minestuck/computer/theme/ComputerTheme.java
@@ -1,12 +1,12 @@
-package com.mraof.minestuck.util;
+package com.mraof.minestuck.computer.theme;
 
 import com.google.gson.*;
 import com.mojang.logging.LogUtils;
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.JsonOps;
+import com.mraof.minestuck.Minestuck;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.util.GsonHelper;
-import net.minecraft.util.StringUtil;
 import net.minecraft.util.valueproviders.ConstantInt;
 import org.slf4j.Logger;
 
@@ -15,6 +15,10 @@ import java.lang.reflect.Type;
 public class ComputerTheme
 {
 	private static final Logger LOGGER = LogUtils.getLogger();
+	
+	public static final String DEFAULT_NAME = "default";
+	public static final int DEFAULT_TEXT_COLOR = 0x404040;
+	public static final ResourceLocation DEFAULT_TEXTURE_PATH = new ResourceLocation(Minestuck.MOD_ID, "textures/gui/theme/default.png");
 	
 	private final ResourceLocation texturePath;
 	private final ConstantInt textColor;
@@ -42,11 +46,6 @@ public class ComputerTheme
 		return themeName;
 	}
 	
-	/*public boolean appliesTo(ResourceLocation texturePathIn)
-	{
-		return texturePathIn == texturePath;
-	}*/
-	
 	public static class Serializer implements JsonDeserializer<ComputerTheme>, JsonSerializer<ComputerTheme>
 	{
 		@Override
@@ -55,7 +54,7 @@ public class ComputerTheme
 			JsonObject json = GsonHelper.convertToJsonObject(jsonElement, "entry");
 			ResourceLocation texturePathProvider = ResourceLocation.CODEC.parse(JsonOps.INSTANCE, json.get("texture_location")).getOrThrow(false, LOGGER::error);
 			ConstantInt textColorProvider = ConstantInt.CODEC.parse(JsonOps.INSTANCE, json.get("text_color")).getOrThrow(false, LOGGER::error);
-			String themeNameProvider = Codec.STRING.parse(JsonOps.INSTANCE, json.get("themeName")).getOrThrow(false, LOGGER::error);
+			String themeNameProvider = Codec.STRING.parse(JsonOps.INSTANCE, json.get("theme_name")).getOrThrow(false, LOGGER::error);
 			return new ComputerTheme(texturePathProvider, textColorProvider, themeNameProvider);
 		}
 		
@@ -65,7 +64,7 @@ public class ComputerTheme
 			JsonObject json = new JsonObject();
 			json.add("texture_location", ResourceLocation.CODEC.encodeStart(JsonOps.INSTANCE, themeData.texturePath).getOrThrow(false, LOGGER::error));
 			json.add("text_color", ConstantInt.CODEC.encodeStart(JsonOps.INSTANCE, themeData.textColor).getOrThrow(false, LOGGER::error));
-			json.add("themeName", Codec.STRING.encodeStart(JsonOps.INSTANCE, themeData.themeName).getOrThrow(false, LOGGER::error));
+			json.add("theme_name", Codec.STRING.encodeStart(JsonOps.INSTANCE, themeData.themeName).getOrThrow(false, LOGGER::error));
 			return json;
 		}
 	}