+++ src/main/java/com/mraof/minestuck/computer/theme/ComputerThemes.java
@@ -1,0 +1,32 @@
+package com.mraof.minestuck.computer.theme;
+
+public enum ComputerThemes
+{
+	/*DEFAULT		(0x404040),
+	PESTERCHUM	(0x404040),
+	TROLLIAN	(0xFF0000),
+	CROCKER		(0x000000),
+	TYPHEUS		(0x6DAFAD),
+	CETUS		(0x9081EE),
+	HEPHAESTUS	(0xFFFFFF),
+	ECHIDNA		(0x005DFF),
+	JOY			(0x282828),
+	SBURB_95 	(0x282828);
+	
+	private final int textColor;
+	
+	ComputerThemes(int textColor)
+	{
+		this.textColor = textColor;
+	}
+	
+	public ComputerThemes next()
+	{
+		return values()[(this.ordinal() + 1) % values().length];
+	}
+	
+	public String getName()
+	{
+		return "theme." + this.name().toLowerCase() + ".name";
+	}*/
+}